+++ ZabbixSender/__init__.py
@@ -1,0 +1,2 @@
+from .ZabbixPacket import ZabbixPacket
+from .ZabbixSender import ZabbixSender