+++ tests/runtests.py
@@ -1,0 +1,7 @@
+import unittest
+
+from test_ZabbixPacket import ZabbixPacketTests
+from test_ZabbixSender import ZabbixSenderTests
+from test_Full import FullTest
+if __name__ == '__main__':
+    unittest.main()